+++ DiskSpaceMonitor.UnitTests/Widgets/WidgetRegistryTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json.Nodes;
 using DiskSpaceMonitor.Widgets;
+using DiskSpaceMonitor.Widgets.Bar;
 using DiskSpaceMonitor.Widgets.Circular;
 using DiskSpaceMonitor.Widgets.Concentric;
 using FluentAssertions;
@@ -67,11 +68,12 @@ namespace DiskSpaceMonitor.UnitTests.Widgets
         [Test]
         public void RealWidgets_ExposeCorrectInstancing()
         {
-            var registry = new WidgetRegistry(new CircularWidget(), new ConcentricWidget());
+            var registry = new WidgetRegistry(new CircularWidget(), new ConcentricWidget(), new BarWidget());
 
-            registry.All.Select(f => f.Id).Should().Equal("Circular", "Concentric");
+            registry.All.Select(f => f.Id).Should().Equal("Circular", "Concentric", "Bar");
             registry.Get("Circular").ShowsAllDrives.Should().BeFalse();
             registry.Get("Concentric").ShowsAllDrives.Should().BeTrue();
+            registry.Get("Bar").ShowsAllDrives.Should().BeTrue();
         }
     }
 }